requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
groq